# farm-market/farm-market-service/src/test/resources/openapi-baseline.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "paths" : {
...
    "/api/v1/auth/me" : {
      "get" : {
        "tags" : [ "Auth" ],
        "operationId" : "me",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserProfile"
                }
              }
            }
          }
        }
      },
      "put" : {
        "tags" : [ "Auth" ],
        "operationId" : "updateMe",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateProfileRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserProfile"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/register" : {
      "post" : {
        "tags" : [ "Auth" ],
        "operationId" : "register",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/refresh" : {
      "post" : {
        "tags" : [ "Auth" ],
        "operationId" : "refresh",
        "parameters" : [ {
          "name" : "farm_refresh",
          "in" : "cookie",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/otp/verify" : {
      "post" : {
        "tags" : [ "Auth" ],
        "operationId" : "otpVerify",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/OtpVerifyRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuthResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/otp/enroll" : {
      "post" : {
        "tags" : [ "Auth" ],
        "operationId" : "otpEnroll",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/OtpEnrollRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/OtpEnrollResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/logout" : {
      "post" : {
        "tags" : [ "Auth" ],
        "operationId" : "logout",
        "parameters" : [ {
          "name" : "farm_refresh",
          "in" : "cookie",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        }
      }
    },
    "/api/v1/auth/login" : {
      "post" : {
        "tags" : [ "Auth" ],
        "operationId" : "login",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/LoginRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/users" : {
      "get" : {
        "tags" : [ "Admin users" ],
        "operationId" : "list_6",
        "parameters" : [ {
          "name" : "role",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "enum" : [ "customer", "staff", "manager", "admin" ]
          }
        }, {
          "name" : "q",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "default" : 20
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PageResponseUserSummary"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "UpdateProfileRequest" : {
        "type" : "object",
        "properties" : {
          "fullName" : {
            "type" : "string",
            "maxLength" : 120,
            "minLength" : 0
          },
          "phone" : {
            "type" : "string",
            "pattern" : "^$|^(\\+84|0)[0-9 .]{8,12}$"
          },
          "birthDate" : {
            "type" : "string",
            "format" : "date"
          },
          "gender" : {
            "type" : "string",
            "pattern" : "^(MALE|FEMALE|OTHER)$"
          }
        },
        "required" : [ "fullName" ]
      },
      "UserProfile" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "email" : {
            "type" : "string"
          },
          "phone" : {
            "type" : "string"
          },
          "fullName" : {
            "type" : "string"
          },
          "birthDate" : {
            "type" : "string",
            "format" : "date"
          },
          "gender" : {
            "type" : "string",
            "enum" : [ "MALE", "FEMALE", "OTHER" ]
          },
          "role" : {
            "type" : "string",
            "enum" : [ "customer", "staff", "manager", "admin" ]
          },
          "branchIds" : {
            "type" : "array",
            "items" : {
              "type" : "integer",
              "format" : "int64"
            }
          }
        }
      },
...
      "UserSummary" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "email" : {
            "type" : "string"
          },
          "fullName" : {
            "type" : "string"
          },
          "role" : {
            "type" : "string",
            "enum" : [ "customer", "staff", "manager", "admin" ]
          },
          "status" : {
            "type" : "string",
            "enum" : [ "ACTIVE", "LOCKED" ]
          },
          "branchIds" : {
            "type" : "array",
            "items" : {
              "type" : "integer",
              "format" : "int64"
            }
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
      "RegisterRequest" : {
        "type" : "object",
        "properties" : {
          "fullName" : {
            "type" : "string",
            "maxLength" : 120,
            "minLength" : 0
          },
          "email" : {
            "type" : "string",
            "format" : "email",
            "maxLength" : 190,
            "minLength" : 0
          },
          "phone" : {
            "type" : "string",
            "pattern" : "^$|^(\\+84|0)[0-9 .]{8,12}$"
          },
          "password" : {
            "type" : "string",
            "maxLength" : 72,
            "minLength" : 8
          }
        },
        "required" : [ "email", "fullName", "password" ]
      },
      "AuthResponse" : {
        "type" : "object",
        "properties" : {
          "accessToken" : {
            "type" : "string"
          },
          "tokenType" : {
            "type" : "string"
          },
          "expiresIn" : {
            "type" : "integer",
            "format" : "int64"
          },
          "user" : {
            "$ref" : "#/components/schemas/UserProfile"
          }
        }
      },
      "OtpVerifyRequest" : {
        "type" : "object",
        "properties" : {
          "challengeToken" : {
            "type" : "string",
            "maxLength" : 2000,
            "minLength" : 0
          },
          "code" : {
            "type" : "string",
            "maxLength" : 10,
            "minLength" : 0
          }
        },
        "required" : [ "challengeToken", "code" ]
      },
      "OtpEnrollRequest" : {
        "type" : "object",
        "properties" : {
          "challengeToken" : {
            "type" : "string",
            "maxLength" : 2000,
            "minLength" : 0
          }
        },
        "required" : [ "challengeToken" ]
      },
      "OtpEnrollResponse" : {
        "type" : "object",
        "properties" : {
          "secret" : {
            "type" : "string"
          },
          "otpauthUri" : {
            "type" : "string"
          }
        }
      },
      "LoginRequest" : {
        "type" : "object",
        "properties" : {
          "identifier" : {
            "type" : "string",
            "maxLength" : 190,
            "minLength" : 0
          },
          "password" : {
            "type" : "string",
            "maxLength" : 72,
            "minLength" : 0
          },
          "clientIp" : {
            "type" : "string",
            "maxLength" : 45,
            "minLength" : 0
          }
        },
        "required" : [ "identifier", "password" ]
      },
...
      "PageMeta" : {
        "type" : "object",
        "properties" : {
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
...
      "PageResponseUserSummary" : {
        "type" : "object",
        "properties" : {
          "data" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/UserSummary"
            }
          },
          "meta" : {
            "$ref" : "#/components/schemas/PageMeta"
          }
        }
      },
...
    }
  }
}